+++ core/src/main/kotlin/no/elg/hex/input/MapEditorInputProcessor.kt
@@ -3,10 +3,8 @@ package no.elg.hex.input
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.Input.Buttons
 import no.elg.hex.hexagon.HexagonData
-import no.elg.hex.hud.MessagesRenderer
 import no.elg.hex.input.editor.EditMetadata
 import no.elg.hex.input.editor.Editor
-import no.elg.hex.island.Island
 import no.elg.hex.screens.MapEditorScreen
 import no.elg.hex.util.calculateHexagonsWithinRadius
 import no.elg.hex.util.coordinates
@@ -42,10 +40,7 @@ class MapEditorInputProcessor(private val screen: MapEditorScreen) : AbstractInp
       editor.edit(hex, data, metadata)
     }
     editor.postEdit(metadata)
-    if (screen.artbSpinner.value != Island.NEVER_PLAYED) {
-      MessagesRenderer.publishWarning("Resetting rounds to beat to unknown as the map has been edited")
-      screen.artbSpinner.setValue(Island.NEVER_PLAYED, true)
-    }
+    screen.resetARtBOnEdit()
   }
 
   private fun setEditorMetadata(cursorHex: Hexagon<HexagonData>): EditMetadata {